Insight_DE/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
Django==3.0.2
djangorestframework==3.11.0
numpy==1.18.1
pandas==1.0.0
pipenv==2018.11.26
psycopg2==2.8.4
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
